+++ lib/features/product/domain/entities/product_entity.dart
@@ -2,7 +2,7 @@ class ProductEntity {
   final int id;
   final String title;
   final String description;
-  final double price;
+  final int price;
   final double discountPercentage;
   final double rating;
   final int stock;
